ATLAS-194: Create module mode

+++ public/js/module.js
@@ -7,32 +7,7 @@ $(function () {
   $("#map_canvas").on("click", "#me-button", function(e) {
     e.preventDefault();
     var id = $(this).val();
-    var site = sites[id].siteData;
-    var auth = {
-      site: site.uuid,
-      token: moduleUUID
-    }
-    var json = JSON.stringify(auth);
-    $.ajax({
-      url: "/api/module/auth?uuid=" + moduleUUID,
-      type: "POST",
-      data: json,
-      dataType: "text",
-    })
-    .done(function(response) {
-      site.module = 1;
-      moduleHasSite = 1;
-      sites[id].siteData = site;
-      sites[id].infowindow.setContent(contentInfowindow(site));
-      sites[id].editwindow.setContent(contentEditwindow(site));
-      repaintMarkers();
-      if(window !== window.top)
-        parent.postMessage("update", "*");
-      bootbox.alert({ message: 'The module is now linked to ' + site.name, backdrop: true });
-    })
-    .fail(function(jqXHR, textStatus, errorThrown) {
-      bootbox.alert({ message: "Error saving your marker - Please try again ! - " + jqXHR.statusText , backdrop: true });
-    });
+    attachMarker(id);
   });
 
   $("#map_canvas").on("click", "#detach-button", function(e) {
@@ -53,10 +28,58 @@ function openBubble(uniqueMarker) {
     google.maps.event.trigger(uniqueMarker, 'click');
 }
 
+function getModuleMarker(module_id, token) {
+  $.ajax({
+      url: "/module",
+      type: "POST",
+      data: encodeURIComponent("module_id")+"="+encodeURIComponent(module_id)+"&"+encodeURIComponent("token")+"="+encodeURIComponent(token),
+  })
+  .done(function(response) {
+    sites[response.id].siteData.module = 1;
+    sites[response.id].infowindow.setContent(contentInfowindow(sites[response.id].siteData));
+  })
+  .fail(function(jqXHR, textStatus, errorThrown) {
+    bootbox.alert({ message: "Error fetching module marker - Please try again ! - " + jqXHR.statusText , backdrop: true });
+  });
+}
+
+function attachMarker(id) {
+  var site = sites[id].siteData;
+  var json = JSON.stringify({ atlas_id: id });
+  $.ajax({
+    url: "/api/module/auth",
+    type: "POST",
+    data: json,
+    dataType: "json",
+    processData: false,
+    contentType: "application/json",
+  })
+  .done(function(response) {
+    Object.keys(sites).forEach(function(siteid) {
+      if(sites[siteid].siteData.module === 1) {
+        delete sites[siteid].siteData.module;
+        sites[siteid].infowindow.setContent(contentInfowindow(sites[siteid].siteData));
+      }
+    });
+    site.module = 1;
+    moduleHasSite = 1;
+    sites[id].siteData = site;
+    sites[id].infowindow.setContent(contentInfowindow(site));
+    repaintMarkers();
+    if(window !== window.top)
+      parent.postMessage("update", "*");
+    bootbox.alert({ message: 'The module is now linked to ' + site.name, backdrop: true });
+  })
+  .fail(function(jqXHR, textStatus, errorThrown) {
+    bootbox.alert({ message: "Error linking module to marker - Please try again ! - " + jqXHR.statusText , backdrop: true });
+  });
+}
+
 function detachMarker(id) {
+  if(!sites[id]) return;
   var site = sites[id].siteData;
   $.ajax({
-    url: "/api/module/auth?uuid=" + moduleUUID,
+    url: "/api/module/auth",
     type: "DELETE",
     dataType: "text",
   })
@@ -65,11 +88,10 @@ function detachMarker(id) {
     moduleHasSite = 0;
     sites[id].siteData = site;
     sites[id].infowindow.setContent(contentInfowindow(site));
-    sites[id].editwindow.setContent(contentEditwindow(site));
     repaintMarkers();
     if(window !== window.top)
       parent.postMessage("update", "*");
-    bootbox.alert({ message: "Authorization delete", backdrop: true });
+    bootbox.alert({ message: "Module detached", backdrop: true });
   })
   .fail(function(jqXHR, textStatus, errorThrown) {
     bootbox.alert({ message: "Error deleting authorization - Please try again ! - " + jqXHR.statusText, backdrop: true });
